refactor(writer): move logic into serializer
