+++ src/JsonApiDotNetCore/Formatters/JsonApiWriter.cs
@@ -2,26 +2,20 @@ using System;
 using System.Text;
 using System.Threading.Tasks;
 using JsonApiDotNetCore.Internal;
-using JsonApiDotNetCore.Models;
 using JsonApiDotNetCore.Serialization;
-using JsonApiDotNetCore.Services;
 using Microsoft.AspNetCore.Mvc.Formatters;
 using Microsoft.Extensions.Logging;
-using Newtonsoft.Json;
 
 namespace JsonApiDotNetCore.Formatters
 {
     public class JsonApiWriter : IJsonApiWriter
     {
         private readonly ILogger<JsonApiWriter> _logger;
-        private readonly IJsonApiContext _jsonApiContext;
         private readonly IJsonApiSerializer _serializer;
 
-        public JsonApiWriter(IJsonApiContext jsonApiContext, 
-            IJsonApiSerializer serializer, 
+        public JsonApiWriter(IJsonApiSerializer serializer, 
             ILoggerFactory loggerFactory)
         {
-            _jsonApiContext = jsonApiContext;
             _serializer = serializer;
             _logger = loggerFactory.CreateLogger<JsonApiWriter>();
         }
@@ -58,36 +52,7 @@ namespace JsonApiDotNetCore.Formatters
 
         private string GetResponseBody(object responseObject)
         {
-            if (responseObject == null)
-                return GetNullDataResponse();
-
-            if (responseObject.GetType() == typeof(Error) || _jsonApiContext.RequestEntity == null)
-                return GetErrorJson(responseObject, _logger);
-            
             return _serializer.Serialize(responseObject);
-        }
-
-        private string GetNullDataResponse()
-        {
-            return JsonConvert.SerializeObject(new Document
-            {
-                Data = null
-            });
-        }
-
-        private string GetErrorJson(object responseObject, ILogger logger)
-        {
-            if (responseObject.GetType() == typeof(Error))
-            {
-                var errors = new ErrorCollection();
-                errors.Add((Error)responseObject);
-                return errors.GetJson();
-            }
-            else
-            {
-                logger?.LogInformation("Response was not a JSONAPI entity. Serializing as plain JSON.");
-                return JsonConvert.SerializeObject(responseObject);
-            }
-        }
+        }        
     }
 }
